Adapt complex d2 models for mapping and children rows
