+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2020-03-23T11:24:10.470Z\n"
-"PO-Revision-Date: 2020-03-23T11:24:10.470Z\n"
+"POT-Creation-Date: 2020-03-24T10:12:11.219Z\n"
+"PO-Revision-Date: 2020-03-24T10:12:11.219Z\n"
 
 msgid "Unauthorized - You do not have permission to view this page."
 msgstr ""
@@ -611,7 +611,13 @@ msgstr ""
 msgid "Importing data in instance {{instance}}"
 msgstr ""
 
-msgid "Program (Indicators)"
+msgid "Data Set with Data Elements"
+msgstr ""
+
+msgid "Program with Data Elements"
+msgstr ""
+
+msgid "Program with Indicators"
 msgstr ""
 
 msgid "Not a valid DHIS2 instance"
